requirements.txt
# HATTO-UFog (Framework) Python dependencies
#
# Recommended: install the Framework package (editable) which pulls its core dependencies
# from `Framework/pyproject.toml`.
#
# Usage:
#   python -m pip install -r requirements.txt
#
-e Framework

# Showcase media rendering helpers (WebGL -> MP4/WebP)
# Note: Playwright still requires a one-time browser install:
#   python -m playwright install chromium
playwright>=1.40,<2.0
imageio-ffmpeg>=0.4.9,<1.0

# AirSim experiments
# Note: AirSim's PythonClient depends on msgpack-rpc-python (which pins tornado<5).
msgpack-rpc-python==0.4.1
backports.ssl_match_hostname==3.7.0.1
msgpack-python==0.5.6

# AirSim run outputs + demo media (MP4/GIF)
numpy>=1.20.0
pandas>=1.3.0,<3.0.0
pyarrow>=10.0.0,<15.0.0
opencv-python>=4.5.0
imageio>=2.25.0
pyyaml>=6.0.0,<7.0.0

# Isaac (postprocess + Framework→Isaac integration helpers)
# Note: Isaac Sim itself is not installed via pip; run Isaac scripts via `$ISAACSIM_ROOT/python.sh`.
# These deps are used by postprocess scripts and integration planning/validation.
# - Isaac/scripts/postprocess_video.py: opencv-python + numpy
# - Isaac/scripts/postprocess_timeseries.py: pandas + pyarrow
# - Isaac/framework_integration/*: pyyaml + numpy

# Optional: plotting / analysis (not required for experiments)
# matplotlib>=3.5,<4.0
# seaborn>=0.11,<1.0
